+++ sdks/js/ReleaseNotes.md
@@ -3,6 +3,12 @@
 
 Included are notes from the latest major revision to current.
 
+## 1.5.7 (13 June, 2016)
+What’s new :
+- Support for standard Video and Ad metadata.
+- API Change: Constants for Standard Video and Ad metadata are available in: ADB.va.plugins.aa.VideoMetadataKeys and ADB.va.plugins.aa.AdMetadataKeys objects.
+- Added video length and video name to the Ad Initiate Adobe Analytics call.
+
 ## 1.5.6 (29 April, 2016)
 What’s new :
 - Added pause tracking feature. Two new heartbeat events are now sent to track pause and stalling.